mod_cladeAnnotator_ui and mod_cladeAnnotator_server

A shiny Module. This module allows the user to add or remove annotations and checks for overlap between those annotations and allows the addition of the heatmap. This module contains 5 functions at the end of the script that are used within (create_tip_list, make_layer, add_map, add_annotations, and current_tree_out) and one function in the golem_utils_server file (snp_anno).

mod_cladeAnnotator_ui(id)

mod_cladeAnnotator_server(
  input,
  output,
  session,
  mFileMatOut,
  make_tree_out,
  add_tree,
  add_anno,
  remove_anno,
  add_heatmap,
  remove_heatmap,
  geneObjectForSNP,
  label_off,
  lab_color,
  mat_off,
  heat_col,
  anno_text,
  median_text
)

Arguments

id
shiny id

input
internal

output
internal

session
internal

mFileMatOut
- from dataDisplay module; is a dataframe

make_tree_out
- from dataDisplay module; class of ggtree, gg, ggplot2

add_tree
- from pushButtons module; is an action button (Y or N)

add_anno
- from pushButtons module; is an action button (Y or N)

remove_anno
- from pushButtons module; is an action button (Y or N)

add_heatmap
- from pushButtons module; is an action button (Y or N)

remove_heatmap
- from pushButtons module; is an action button (Y or N)

geneObjectForSNP
- from upload/example Data module. Is a tibble

lab_color
- from paramsTree; select input from user (multiple choice)

mat_off
- from paramsTree module; this is numeric input from user

anno_text
- from paramsTree; user input text (blank or any word)

median_text
- from paramsTree; user input text (blank or any word)

label_off-
from paramsTree module; this is numeric input from user

heat_color
- from paramsTree; select input from user (multiple choice)

Details

cladeAnnotator Function